Remove unnecessary test abstraction
There's only one implementation of the base test class, so
merge them into a single one.
